MGRENTITLE-20 Implement upgrade operation for Kong routes (#73)
- Refactor module flow to perform individually for upgrade/entitle/revoke operations.
- Implement different flows for UI and BE modules
- Adjust kong/keycloak resource creator stages for Folio flow to execute per module
- Fix unit tests
- Fix integration tests

+++ src/main/java/org/folio/entitlement/integration/keycloak/KeycloakAuthResourceCreator.java
@@ -3,10 +3,12 @@ package org.folio.entitlement.integration.keycloak;
 import static org.apache.commons.collections4.ListUtils.emptyIfNull;
 
 import lombok.RequiredArgsConstructor;
-import org.folio.entitlement.integration.folio.ApplicationStageContext;
+import lombok.extern.log4j.Log4j2;
+import org.folio.entitlement.domain.model.ApplicationStageContext;
 import org.folio.entitlement.service.stage.DatabaseLoggingStage;
 import org.keycloak.admin.client.Keycloak;
 
+@Log4j2
 @RequiredArgsConstructor
 public class KeycloakAuthResourceCreator extends DatabaseLoggingStage<ApplicationStageContext> {
 
@@ -14,19 +16,26 @@ public class KeycloakAuthResourceCreator extends DatabaseLoggingStage<Applicatio
   private final KeycloakService keycloakService;
 
   @Override
-  public void execute(ApplicationStageContext context) {
-    var realm = context.getTenantName();
-    var applicationDescriptor = context.getApplicationDescriptor();
+  public void execute(ApplicationStageContext ctx) {
+    var realm = ctx.getTenantName();
+    var applicationDescriptor = ctx.getApplicationDescriptor();
     var moduleDescriptors = applicationDescriptor.getModuleDescriptors();
     keycloakClient.tokenManager().grantToken();
     emptyIfNull(moduleDescriptors).forEach(desc -> keycloakService.updateAuthResources(null, desc, realm));
   }
 
   @Override
-  public void cancel(ApplicationStageContext context) {
-    var realm = context.getTenantName();
-    var applicationDescriptor = context.getApplicationDescriptor();
+  public void cancel(ApplicationStageContext ctx) {
+    var request = ctx.getEntitlementRequest();
+    if (!request.isPurgeOnRollback()) {
+      log.debug("Skipping purge of Keycloak resource during rollback: applicationId = {}",  ctx.getApplicationId());
+      return;
+    }
+
+    var realm = ctx.getTenantName();
+    var applicationDescriptor = ctx.getApplicationDescriptor();
     var moduleDescriptors = applicationDescriptor.getModuleDescriptors();
+
     keycloakClient.tokenManager().grantToken();
     emptyIfNull(moduleDescriptors).forEach(descriptor -> keycloakService.removeAuthResources(descriptor, realm));
   }
