+++ src/main/java/org/folio/entitlement/integration/kong/KongRouteCreator.java
@@ -1,10 +1,12 @@
 package org.folio.entitlement.integration.kong;
 
 import lombok.RequiredArgsConstructor;
-import org.folio.entitlement.integration.folio.ApplicationStageContext;
+import lombok.extern.log4j.Log4j2;
+import org.folio.entitlement.domain.model.ApplicationStageContext;
 import org.folio.entitlement.service.stage.DatabaseLoggingStage;
 import org.folio.tools.kong.service.KongGatewayService;
 
+@Log4j2
 @RequiredArgsConstructor
 public class KongRouteCreator extends DatabaseLoggingStage<ApplicationStageContext> {
 
@@ -19,6 +21,12 @@ public class KongRouteCreator extends DatabaseLoggingStage<ApplicationStageConte
 
   @Override
   public void cancel(ApplicationStageContext context) {
+    var request = context.getEntitlementRequest();
+    if (!request.isPurgeOnRollback()) {
+      log.debug("Skipping purge of Kong routes during rollback: applicationId = {}",  context.getApplicationId());
+      return;
+    }
+
     var tenantName = context.getTenantName();
     var applicationDescriptor = context.getApplicationDescriptor();
     kongGatewayService.removeRoutes(tenantName, applicationDescriptor.getModuleDescriptors());
